+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/impl/ContractSynchronizationExecutor.java
@@ -28,11 +28,13 @@ import eu.bcvsolutions.idm.acc.dto.AccContractAccountDto;
 import eu.bcvsolutions.idm.acc.dto.EntityAccountDto;
 import eu.bcvsolutions.idm.acc.dto.SyncIdentityContractDto;
 import eu.bcvsolutions.idm.acc.dto.SysSyncActionLogDto;
+import eu.bcvsolutions.idm.acc.dto.SysSyncContractConfigDto;
 import eu.bcvsolutions.idm.acc.dto.SysSyncItemLogDto;
 import eu.bcvsolutions.idm.acc.dto.SysSyncLogDto;
 import eu.bcvsolutions.idm.acc.dto.SysSystemAttributeMappingDto;
 import eu.bcvsolutions.idm.acc.dto.filter.AccContractAccountFilter;
 import eu.bcvsolutions.idm.acc.dto.filter.EntityAccountFilter;
+import eu.bcvsolutions.idm.acc.entity.SysSyncContractConfig_;
 import eu.bcvsolutions.idm.acc.exception.ProvisioningException;
 import eu.bcvsolutions.idm.acc.service.api.AccAccountService;
 import eu.bcvsolutions.idm.acc.service.api.AccContractAccountService;
@@ -48,24 +50,31 @@ import eu.bcvsolutions.idm.acc.service.api.SysSystemAttributeMappingService;
 import eu.bcvsolutions.idm.acc.service.api.SysSystemEntityService;
 import eu.bcvsolutions.idm.acc.service.api.SysSystemMappingService;
 import eu.bcvsolutions.idm.acc.service.api.SysSystemService;
+import eu.bcvsolutions.idm.core.api.domain.ContractState;
 import eu.bcvsolutions.idm.core.api.dto.IdmContractGuaranteeDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityContractDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
+import eu.bcvsolutions.idm.core.api.dto.IdmTreeNodeDto;
+import eu.bcvsolutions.idm.core.api.dto.IdmTreeTypeDto;
 import eu.bcvsolutions.idm.core.api.dto.filter.CorrelationFilter;
 import eu.bcvsolutions.idm.core.api.dto.filter.IdmContractGuaranteeFilter;
 import eu.bcvsolutions.idm.core.api.dto.filter.IdmIdentityContractFilter;
+import eu.bcvsolutions.idm.core.api.dto.filter.IdmTreeNodeFilter;
 import eu.bcvsolutions.idm.core.api.event.EntityEvent;
 import eu.bcvsolutions.idm.core.api.service.ConfidentialStorage;
 import eu.bcvsolutions.idm.core.api.service.EntityEventManager;
 import eu.bcvsolutions.idm.core.api.service.GroovyScriptService;
 import eu.bcvsolutions.idm.core.api.service.IdmContractGuaranteeService;
 import eu.bcvsolutions.idm.core.api.service.IdmIdentityContractService;
+import eu.bcvsolutions.idm.core.api.service.IdmTreeNodeService;
 import eu.bcvsolutions.idm.core.api.service.LookupService;
 import eu.bcvsolutions.idm.core.api.service.ReadWriteDtoService;
+import eu.bcvsolutions.idm.core.api.utils.DtoUtils;
 import eu.bcvsolutions.idm.core.api.utils.EntityUtils;
 import eu.bcvsolutions.idm.core.eav.api.entity.FormableEntity;
 import eu.bcvsolutions.idm.core.eav.api.service.FormService;
 import eu.bcvsolutions.idm.core.model.entity.IdmIdentityContract;
+import eu.bcvsolutions.idm.core.model.entity.IdmTreeNode_;
 import eu.bcvsolutions.idm.core.model.event.ContractGuaranteeEvent;
 import eu.bcvsolutions.idm.core.model.event.ContractGuaranteeEvent.ContractGuaranteeEventType;
 import eu.bcvsolutions.idm.core.model.event.IdentityContractEvent;
@@ -81,10 +90,12 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 	private final IdmIdentityContractService contractService;
 	private final AccContractAccountService contractAccoutnService;
 	private final IdmContractGuaranteeService guaranteeService;
+	private final IdmTreeNodeService treeNodeService;
 	private final LookupService lookupService;
 	public final static String CONTRACT_STATE_FIELD = "state";
 	public final static String CONTRACT_GUARANTEES_FIELD = "guarantees";
 	public final static String CONTRACT_IDENTITY_FIELD = "identity";
+	public final static String CONTRACT_WORK_POSITION_FIELD = "workPosition";
 	public final static String SYNC_CONTRACT_FIELD = "sync_contract";
 
 	@Autowired
@@ -99,7 +110,7 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 			WorkflowProcessInstanceService workflowProcessInstanceService, EntityManager entityManager,
 			SysSystemMappingService systemMappingService, SysSchemaObjectClassService schemaObjectClassService,
 			SysSchemaAttributeService schemaAttributeService, LookupService lookupService,
-			IdmContractGuaranteeService guaranteeService) {
+			IdmContractGuaranteeService guaranteeService, IdmTreeNodeService treeNodeService) {
 		super(connectorFacade, systemService, attributeHandlingService, synchronizationConfigService,
 				synchronizationLogService, syncActionLogService, accountService, systemEntityService,
 				confidentialStorage, formService, syncItemLogService, entityEventManager, groovyScriptService,
@@ -110,11 +121,13 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 		Assert.notNull(contractAccoutnService, "Contract-account service is mandatory!");
 		Assert.notNull(lookupService, "Lookup service is mandatory!");
 		Assert.notNull(guaranteeService, "Contract guarantee service is mandatory!");
+		Assert.notNull(treeNodeService, "Tree node service is mandatory!");
 		//
 		this.contractService = contractService;
 		this.contractAccoutnService = contractAccoutnService;
 		this.lookupService = lookupService;
 		this.guaranteeService = guaranteeService;
+		this.treeNodeService = treeNodeService;
 	}
 
 	/**
@@ -153,10 +166,9 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 	@Override
 	protected void callProvisioningForEntity(IdmIdentityContractDto entity, SystemEntityType entityType,
 			SysSyncItemLogDto logItem) {
-		addToItemLog(logItem,
-				MessageFormat.format(
-						"Call provisioning (process IdentityContractEvent.UPDATE) for contract ({0}) with position ({1}).",
-						entity.getId(), entity.getPosition()));
+		addToItemLog(logItem, MessageFormat.format(
+				"Call provisioning (process IdentityContractEvent.UPDATE) for contract ({0}) with position ({1}).",
+				entity.getId(), entity.getPosition()));
 		entityEventManager.process(new IdentityContractEvent(IdentityContractEventType.UPDATE, entity)).getContent();
 	}
 
@@ -188,10 +200,9 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 			// We will remove contract account, but without delete connected
 			// account
 			contractAccoutnService.delete(entityAccount, false);
-			addToItemLog(logItem,
-					MessageFormat.format(
-							"Contract-account relation deleted (without call delete provisioning) (contract id: {0}, contract-account id: {1})",
-							entityAccount.getContract(), entityAccount.getId()));
+			addToItemLog(logItem, MessageFormat.format(
+					"Contract-account relation deleted (without call delete provisioning) (contract id: {0}, contract-account id: {1})",
+					entityAccount.getContract(), entityAccount.getId()));
 
 		});
 		return;
@@ -230,9 +241,12 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 			Object transformedValue = getValueByMappedAttribute(attribute, icAttributes, context);
 			// Guarantees will be set no to the dto (we does not have field for
 			// they), but to the embedded map.
-			if (CONTRACT_GUARANTEES_FIELD.equals(attributeProperty)
-					&& transformedValue instanceof SyncIdentityContractDto) {
-				dto.getEmbedded().put(SYNC_CONTRACT_FIELD, (SyncIdentityContractDto) transformedValue);
+			if (CONTRACT_GUARANTEES_FIELD.equals(attributeProperty)) {
+				if (transformedValue instanceof SyncIdentityContractDto) {
+					dto.getEmbedded().put(SYNC_CONTRACT_FIELD, (SyncIdentityContractDto) transformedValue);
+				} else {
+					dto.getEmbedded().put(SYNC_CONTRACT_FIELD, new SyncIdentityContractDto());
+				}
 				return;
 			}
 			// Set transformed value from target system to entity
@@ -255,12 +269,51 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 		// Transform contract state enumeration from string
 		if (CONTRACT_STATE_FIELD.equals(attribute.getIdmPropertyName()) && transformedValue instanceof String
 				&& attribute.isEntityAttribute()) {
-			// transformedValue = IdentityContractType.valueOf((String)
-			// transformedValue);
+			return ContractState.valueOf((String) transformedValue);
 		}
 		// Transform contract guarantees
-		if (transformedValue != null && CONTRACT_GUARANTEES_FIELD.equals(attribute.getIdmPropertyName())
+		if (CONTRACT_GUARANTEES_FIELD.equals(attribute.getIdmPropertyName()) && attribute.isEntityAttribute()) {
+			return transformGuarantees(context, transformedValue);
+		}
+		// Transform work position (tree node)
+		if (CONTRACT_WORK_POSITION_FIELD.equals(attribute.getIdmPropertyName()) && attribute.isEntityAttribute()) {
+
+			if (transformedValue != null) {
+				IdmTreeNodeDto workposition = this.findTreeNode(transformedValue, context);
+				if (workposition != null) {
+					return workposition.getId();
+				}
+				return null;
+			} else {
+				if (getConfig(context).getDefaultTreeNode() != null) {
+					UUID defaultNode = ((SysSyncContractConfigDto) context.getConfig()).getDefaultTreeNode();
+					IdmTreeNodeDto node = (IdmTreeNodeDto) lookupService.lookupDto(IdmTreeNodeDto.class, defaultNode);
+					if (node != null) {
+						context.getLogItem().addToLog(MessageFormat.format(
+								"Warning! - None workposition was defined for this realtion, we use default workposition [{0}]!",
+								node.getCode()));
+						return node.getId();
+					}
+				}
+			}
+		}
+		// Transform contract owner
+		if (transformedValue != null && CONTRACT_IDENTITY_FIELD.equals(attribute.getIdmPropertyName())
 				&& attribute.isEntityAttribute()) {
+			context.getLogItem().addToLog(MessageFormat.format("Finding contract owner [{0}].", transformedValue));
+			IdmIdentityDto identity = this.findIdentity(transformedValue, context);
+			if (identity == null) {
+				throw new ProvisioningException(AccResultCode.SYNCHRONIZATION_IDM_FIELD_CANNOT_BE_NULL,
+						ImmutableMap.of("property", CONTRACT_IDENTITY_FIELD));
+			}
+			return identity.getId();
+		}
+
+		return transformedValue;
+	}
+
+	private Object transformGuarantees(SynchronizationContext context, Object transformedValue) {
+		if (transformedValue != null) {
 			SyncIdentityContractDto syncContract = new SyncIdentityContractDto();
 			if (transformedValue instanceof List) {
 				((List<?>) transformedValue).stream().forEach(guarantee -> {
@@ -283,19 +336,21 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 				}
 			}
 			transformedValue = syncContract;
-		}
-		// Transform contract owner
-		if (transformedValue != null && CONTRACT_IDENTITY_FIELD.equals(attribute.getIdmPropertyName())
-				&& attribute.isEntityAttribute()) {
-			context.getLogItem().addToLog(MessageFormat.format("Finding contract owner [{0}].", transformedValue));
-			IdmIdentityDto identity = this.findIdentity(transformedValue, context);
-			if(identity == null){
-				throw new ProvisioningException(AccResultCode.SYNCHRONIZATION_IDM_FIELD_CANNOT_BE_NULL,
-						ImmutableMap.of("property", CONTRACT_IDENTITY_FIELD));
+		} else {
+			if (getConfig(context).getDefaultLeader() != null) {
+				UUID defaultLeader = ((SysSyncContractConfigDto) context.getConfig()).getDefaultLeader();
+				IdmIdentityDto identity = (IdmIdentityDto) lookupService.lookupDto(IdmIdentityDto.class, defaultLeader);
+				if (identity != null) {
+					SyncIdentityContractDto syncContract = new SyncIdentityContractDto();
+					syncContract.getGuarantees().add(identity);
+					transformedValue = syncContract;
+					context.getLogItem()
+							.addToLog(MessageFormat.format(
+									"Warning! - None leader was found for this realtion, we use default leader [{0}]!",
+									identity.getCode()));
+				}
 			}
-			transformedValue = identity.getId();
 		}
-
 		return transformedValue;
 	}
 
@@ -305,8 +360,8 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 					(Serializable) value);
 
 			if (identity == null) {
-				context.getLogItem().addToLog(MessageFormat.format(
-						"Warning! - Identity [{0}] was not found for [{0}]!", value));
+				context.getLogItem()
+						.addToLog(MessageFormat.format("Warning! - Identity [{0}] was not found for [{0}]!", value));
 				this.initSyncActionLog(context.getActionType(), OperationResultType.WARNING, context.getLogItem(),
 						context.getLog(), context.getActionLogs());
 				return null;
@@ -322,6 +377,70 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 					context.getLog(), context.getActionLogs());
 		}
 		return null;
+	}
+
+	private IdmTreeNodeDto findTreeNode(Object value, SynchronizationContext context) {
+		if (value instanceof Serializable) {
+			// Find by UUID
+			context.getLogItem().addToLog(
+					MessageFormat.format("Work position - try find directly by transformed value [{0}]!", value));
+			IdmTreeNodeDto node = (IdmTreeNodeDto) lookupService.lookupDto(IdmTreeNodeDto.class, (Serializable) value);
+
+			if (node != null) {
+				IdmTreeTypeDto treeTypeDto = DtoUtils.getEmbedded(node, IdmTreeNode_.treeType, IdmTreeTypeDto.class);
+				context.getLogItem().addToLog(MessageFormat.format(
+						"Work position - One node [{1}] (in tree type [{2}]) was found directly by transformed value [{0}]!",
+						value, node.getCode(), treeTypeDto.getCode()));
+				return node;
+			}
+			context.getLogItem().addToLog(MessageFormat
+					.format("Work position - was not not found directly from transformed value [{0}]!", value));
+			if (value instanceof String) {
+				// Find by code in default tree type
+				IdmTreeNodeFilter treeNodeFilter = new IdmTreeNodeFilter();
+				SysSyncContractConfigDto config = this.getConfig(context);
+				IdmTreeTypeDto defaultTreeType = DtoUtils.getEmbedded(config, SysSyncContractConfig_.defaultTreeType,
+						IdmTreeTypeDto.class);
+				treeNodeFilter.setTreeTypeId(config.getDefaultTreeType());
+				treeNodeFilter.setCode((String) value);
+				context.getLogItem()
+						.addToLog(MessageFormat.format(
+								"Work position - try find in default tree type [{1}] with code [{0}]!", value,
+								defaultTreeType.getCode()));
+				List<IdmTreeNodeDto> nodes = treeNodeService.find(treeNodeFilter, null).getContent();
+				if (nodes.isEmpty()) {
+					context.getLogItem().addToLog(
+							MessageFormat.format("Warning - Work position - none node found for code [{0}]!", value));
+					this.initSyncActionLog(context.getActionType(), OperationResultType.WARNING, context.getLogItem(),
+							context.getLog(), context.getActionLogs());
+					return null;
+				} else if (nodes.size() > 1) {
+					context.getLogItem()
+							.addToLog(MessageFormat.format(
+									"Warning - Work position - more then one [{0}] node found for code [{1}]!", value,
+									nodes.size()));
+					this.initSyncActionLog(context.getActionType(), OperationResultType.WARNING, context.getLogItem(),
+							context.getLog(), context.getActionLogs());
+					return null;
+
+				} else {
+					MessageFormat.format("Work position - One node [{1}] was found for code [{0}]!", value,
+							nodes.get(0).getId());
+					return nodes.get(0);
+				}
+			}
+		} else {
+			context.getLogItem().addToLog(MessageFormat.format(
+					"Warning! - Work position cannot be found, because transformed value [{0}] is not Serializable!",
+					value));
+			this.initSyncActionLog(context.getActionType(), OperationResultType.WARNING, context.getLogItem(),
+					context.getLog(), context.getActionLogs());
+		}
+		return null;
+	}
+
+	private SysSyncContractConfigDto getConfig(SynchronizationContext context) {
+		return ((SysSyncContractConfigDto) context.getConfig());
 	}
 
 	/**
@@ -345,7 +464,7 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 			guaranteeFilter.setIdentityContractId(entity.getId());
 
 			List<IdmContractGuaranteeDto> currentGuarantees = guaranteeService.find(guaranteeFilter, null).getContent();
-			
+
 			// Search guarantees to delete
 			List<IdmContractGuaranteeDto> guaranteesToDelete = currentGuarantees.stream().filter(sysImplementer -> {
 				return sysImplementer.getGuarantee() != null
@@ -358,7 +477,7 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 					return identity.getId().equals(currentGuarrantee.getGuarantee());
 				}).findFirst().isPresent();
 			}).collect(Collectors.toList());
-			
+
 			// Delete guarantees
 			guaranteesToDelete.forEach(guarantee -> {
 				EntityEvent<IdmContractGuaranteeDto> guaranteeEvent = new ContractGuaranteeEvent(
@@ -366,7 +485,7 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 						ImmutableMap.of(ProvisioningService.SKIP_PROVISIONING, skipProvisioning));
 				guaranteeService.publish(guaranteeEvent);
 			});
-			
+
 			// Create new guarantees
 			guaranteesToAdd.forEach(identity -> {
 				IdmContractGuaranteeDto guarantee = new IdmContractGuaranteeDto();