+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/impl/DefaultSysSyncConfigService.java
@@ -9,15 +9,23 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
 
+import com.google.common.collect.ImmutableMap;
+
+import eu.bcvsolutions.idm.acc.domain.AccResultCode;
+import eu.bcvsolutions.idm.acc.dto.AbstractSysSyncConfigDto;
 import eu.bcvsolutions.idm.acc.dto.SysSyncConfigDto;
+import eu.bcvsolutions.idm.acc.dto.SysSyncContractConfigDto;
 import eu.bcvsolutions.idm.acc.dto.SysSystemMappingDto;
 import eu.bcvsolutions.idm.acc.dto.filter.SysSyncConfigFilter;
 import eu.bcvsolutions.idm.acc.dto.filter.SysSyncLogFilter;
 import eu.bcvsolutions.idm.acc.entity.SysSyncConfig;
+import eu.bcvsolutions.idm.acc.entity.SysSyncContractConfig;
 import eu.bcvsolutions.idm.acc.repository.SysSyncConfigRepository;
 import eu.bcvsolutions.idm.acc.service.api.SysSyncConfigService;
 import eu.bcvsolutions.idm.acc.service.api.SysSyncLogService;
+import eu.bcvsolutions.idm.core.api.exception.ResultCodeException;
 import eu.bcvsolutions.idm.core.api.service.AbstractReadWriteDtoService;
+import eu.bcvsolutions.idm.core.api.service.LookupService;
 import eu.bcvsolutions.idm.core.api.utils.EntityUtils;
 import eu.bcvsolutions.idm.core.security.api.domain.BasePermission;
 
@@ -29,25 +37,42 @@ import eu.bcvsolutions.idm.core.security.api.domain.BasePermission;
  */
 @Service
 public class DefaultSysSyncConfigService
-		extends AbstractReadWriteDtoService<SysSyncConfigDto, SysSyncConfig, SysSyncConfigFilter>
+		extends AbstractReadWriteDtoService<AbstractSysSyncConfigDto, SysSyncConfig, SysSyncConfigFilter>
 		implements SysSyncConfigService {
 
 	private final SysSyncConfigRepository repository;
 	private final SysSyncLogService synchronizationLogService;
+	private final LookupService lookupService;
 
 	@Autowired
-	public DefaultSysSyncConfigService(SysSyncConfigRepository repository,
-			SysSyncLogService synchronizationLogService) {
+	public DefaultSysSyncConfigService(SysSyncConfigRepository repository, SysSyncLogService synchronizationLogService,
+			LookupService lookupService) {
 		super(repository);
 		//
 		Assert.notNull(synchronizationLogService);
+		Assert.notNull(lookupService);
 		//
 		this.repository = repository;
 		this.synchronizationLogService = synchronizationLogService;
+		this.lookupService = lookupService;
 	}
-	
+
 	@Override
-	protected Page<SysSyncConfig> findEntities(SysSyncConfigFilter filter, Pageable pageable, BasePermission... permission) {
+	public AbstractSysSyncConfigDto save(AbstractSysSyncConfigDto dto, BasePermission... permission) {
+		if (dto != null && !this.isNew(dto)) {
+			AbstractSysSyncConfigDto persistedConfig = this.get(dto.getId(), permission);
+			if (!dto.getClass().equals(persistedConfig.getClass())) {
+				throw new ResultCodeException(AccResultCode.SYNCHRONIZATION_CONFIG_TYPE_CANNOT_BE_CANGED,
+						ImmutableMap.of("old", persistedConfig.getClass().getSimpleName(), "new",
+								persistedConfig.getClass().getSimpleName()));
+			}
+		}
+		return super.save(dto, permission);
+	}
+
+	@Override
+	protected Page<SysSyncConfig> findEntities(SysSyncConfigFilter filter, Pageable pageable,
+			BasePermission... permission) {
 		if (filter == null) {
 			return repository.findAll(pageable);
 		}
@@ -55,8 +80,24 @@ public class DefaultSysSyncConfigService
 	}
 
 	@Override
+	protected Class<? extends SysSyncConfig> getEntityClass(AbstractSysSyncConfigDto dto) {
+		if (dto instanceof SysSyncContractConfigDto) {
+			return SysSyncContractConfig.class;
+		}
+		return SysSyncConfig.class;
+	}
+
+	@Override
+	protected Class<? extends AbstractSysSyncConfigDto> getDtoClass(SysSyncConfig entity) {
+		if (entity instanceof SysSyncContractConfig) {
+			return SysSyncContractConfigDto.class;
+		}
+		return SysSyncConfigDto.class;
+	}
+
+	@Override
 	@Transactional
-	public void delete(SysSyncConfigDto synchronizationConfig, BasePermission... permission) {
+	public void delete(AbstractSysSyncConfigDto synchronizationConfig, BasePermission... permission) {
 		Assert.notNull(synchronizationConfig);
 		checkAccess(getEntity(synchronizationConfig.getId()), permission);
 		//
@@ -69,22 +110,22 @@ public class DefaultSysSyncConfigService
 		//
 		super.delete(synchronizationConfig);
 	}
-	
+
 	@Override
-	public boolean isRunning(SysSyncConfigDto config){
-		if(config == null){
+	public boolean isRunning(AbstractSysSyncConfigDto config) {
+		if (config == null) {
 			return false;
 		}
 		int count = ((SysSyncConfigRepository) this.getRepository())
 				.runningCount(((SysSyncConfigRepository) this.getRepository()).findOne(config.getId()));
 		return count > 0;
 	}
-	
+
 	@Override
-	public SysSyncConfigDto clone(UUID id) {
-		SysSyncConfigDto original = this.get(id);
+	public AbstractSysSyncConfigDto clone(UUID id) {
+		AbstractSysSyncConfigDto original = this.get(id);
 		Assert.notNull(original, "Config of synchronization must be found!");
-		
+
 		// We do detach this entity (and set id to null)
 		original.setId(null);
 		EntityUtils.clearAuditFields(original);
@@ -97,6 +138,5 @@ public class DefaultSysSyncConfigService
 		Assert.notNull(mappingDto.getId());
 		return repository.countByCorrelationAttribute_Id(mappingDto.getId());
 	}
-
 
 }