+++ Realization/backend/ic/src/main/java/eu/bcvsolutions/idm/ic/czechidm/service/impl/CzechIdMIcConnectorService.java
@@ -1,9 +1,7 @@
 package eu.bcvsolutions.idm.ic.czechidm.service.impl;
 
 import java.text.MessageFormat;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;