+++ Realization/frontend/czechidm-acc/src/services/ContractAccountService.js
@@ -1,0 +1,43 @@
+import { Services } from 'czechidm-core';
+import { Domain } from 'czechidm-core';
+import AccountTypeEnum from '../domain/AccountTypeEnum';
+
+
+/**
+ * Contract accounts
+ *
+ * @author Vít Švanda
+ */
+export default class ContractAccountService extends Services.AbstractService {
+
+  constructor() {
+    super();
+  }
+
+  getNiceLabel(entity) {
+    if (!entity) {
+      return '';
+    }
+    return `${AccountTypeEnum.getNiceLabel(entity._embedded.account.accountType)}:${entity._embedded.account._embedded.system.name}:${entity._embedded.account.uid}`;
+  }
+
+  getApiPath() {
+    return '/contract-accounts';
+  }
+
+  supportsPatch() {
+    return false;
+  }
+
+  supportsAuthorization() {
+    return true;
+  }
+
+  getGroupPermission() {
+    return 'CONTRACTACCOUNT';
+  }
+
+  getDefaultSearchParameters() {
+    return super.getDefaultSearchParameters().setName(Domain.SearchParameters.NAME_QUICK).clearSort().setSort('account.uid');
+  }
+}